download html detect and convert encoding to utf-8
in original version, download document would be utf-8,
but the charset meta tag will not be fix.

current version will change `<meta charset>` and
`<meta http-equiv>` to utf-8, and the original encoding
will moved to `<meta data-gholk-original-charset>`
or `<meta data-gholk-original-content>`.

if no charset tag detected, this script will ask you whether
add a new charset tag or leave it as if.